requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Pillow==10.2.0
numpy==1.26.4
scikit-image==0.23.2
scikit-learn==1.4.1.post1
joblib==1.3.2
Flask-Limiter==3.11.0
exifread==3.3.1 
gunicorn==23.0.0